adjust docs examples test to new docs structure
